REFACTOR #164 : image 패키지 구조 리팩토링
