+++ src/main/java/com/lokoko/domain/image/domain/repository/ReviewImageRepository.java
@@ -1,6 +1,6 @@
-package com.lokoko.domain.image.repository;
+package com.lokoko.domain.image.domain.repository;
 
-import com.lokoko.domain.image.entity.ReviewImage;
+import com.lokoko.domain.image.domain.entity.ReviewImage;
 import com.lokoko.domain.review.domain.entity.Review;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
