+++ src/main/java/com/lokoko/domain/review/application/service/ReviewDetailsService.java
@@ -1,11 +1,11 @@
 package com.lokoko.domain.review.application.service;
 
-import com.lokoko.domain.image.entity.ProductImage;
-import com.lokoko.domain.image.entity.ReceiptImage;
-import com.lokoko.domain.image.entity.ReviewImage;
-import com.lokoko.domain.image.repository.ProductImageRepository;
-import com.lokoko.domain.image.repository.ReceiptImageRepository;
-import com.lokoko.domain.image.repository.ReviewImageRepository;
+import com.lokoko.domain.image.domain.entity.ProductImage;
+import com.lokoko.domain.image.domain.entity.ReceiptImage;
+import com.lokoko.domain.image.domain.entity.ReviewImage;
+import com.lokoko.domain.image.domain.repository.ProductImageRepository;
+import com.lokoko.domain.image.domain.repository.ReceiptImageRepository;
+import com.lokoko.domain.image.domain.repository.ReviewImageRepository;
 import com.lokoko.domain.like.domain.repository.ReviewLikeRepository;
 import com.lokoko.domain.product.domain.entity.Product;
 import com.lokoko.domain.review.api.dto.response.ImageReviewDetailResponse;
@@ -15,10 +15,10 @@ import com.lokoko.domain.review.exception.ProductImageNotFoundException;
 import com.lokoko.domain.review.exception.ReviewNotFoundException;
 import com.lokoko.domain.review.exception.ReviewVideoNotFoundException;
 import com.lokoko.domain.review.domain.repository.ReviewRepository;
-import com.lokoko.domain.user.entity.User;
-import com.lokoko.domain.user.entity.enums.Role;
+import com.lokoko.domain.user.domain.entity.User;
+import com.lokoko.domain.user.domain.entity.enums.Role;
 import com.lokoko.domain.user.exception.UserNotFoundException;
-import com.lokoko.domain.user.repository.UserRepository;
+import com.lokoko.domain.user.domain.repository.UserRepository;
 import com.lokoko.domain.video.domain.entity.ReviewVideo;
 import com.lokoko.domain.video.domain.repository.ReviewVideoRepository;
 import java.util.List;
