+++ src/main/java/com/lokoko/domain/review/application/service/ReviewService.java
@@ -3,10 +3,10 @@ package com.lokoko.domain.review.application.service;
 
 import static com.lokoko.global.utils.AllowedMediaType.ALLOWED_MEDIA_TYPES;
 
-import com.lokoko.domain.image.entity.ReceiptImage;
-import com.lokoko.domain.image.entity.ReviewImage;
-import com.lokoko.domain.image.repository.ReceiptImageRepository;
-import com.lokoko.domain.image.repository.ReviewImageRepository;
+import com.lokoko.domain.image.domain.entity.ReceiptImage;
+import com.lokoko.domain.image.domain.entity.ReviewImage;
+import com.lokoko.domain.image.domain.repository.ReceiptImageRepository;
+import com.lokoko.domain.image.domain.repository.ReviewImageRepository;
 import com.lokoko.domain.like.domain.repository.ReviewLikeRepository;
 import com.lokoko.domain.product.domain.entity.Product;
 import com.lokoko.domain.product.domain.entity.ProductOption;
@@ -37,10 +37,10 @@ import com.lokoko.domain.review.exception.ReviewNotFoundException;
 import com.lokoko.domain.review.exception.ReviewPermissionException;
 import com.lokoko.domain.review.mapper.ReviewMapper;
 import com.lokoko.domain.review.domain.repository.ReviewRepository;
-import com.lokoko.domain.user.entity.User;
-import com.lokoko.domain.user.entity.enums.Role;
+import com.lokoko.domain.user.domain.entity.User;
+import com.lokoko.domain.user.domain.entity.enums.Role;
 import com.lokoko.domain.user.exception.UserNotFoundException;
-import com.lokoko.domain.user.repository.UserRepository;
+import com.lokoko.domain.user.domain.repository.UserRepository;
 import com.lokoko.domain.video.domain.entity.ReviewVideo;
 import com.lokoko.domain.video.domain.repository.ReviewVideoRepository;
 import com.lokoko.global.common.dto.PresignedUrlResponse;
